changings in categories and fixing bugs

+++ AngulaToDo.Server/Repositories/CategoryRepository.cs
@@ -17,6 +17,7 @@ namespace AngulaToDo.Server.Repositories
         public async Task<IEnumerable<Category>> GetAllCategoriesAsync(string userId)
         {
             var categories = await _context.Categories
+                .AsNoTracking()
                 .Where(c => c.UserId == userId)
                 .ToListAsync();
 
@@ -26,34 +27,33 @@ namespace AngulaToDo.Server.Repositories
         public async Task<Category> GetCategoryByNameAsync(string userId, string name)
         {
             var category = await _context.Categories
+                .AsNoTracking()
                 .FirstOrDefaultAsync(c => c.Name == name && c.UserId == userId);
 
             return category;
         }
 
-        public async Task<Category> CreateCategoryAsync(string userId, CategoryDto categoryDto)
+        public async Task<Category> CreateCategoryAsync(string userId, Category category)
         {
-            var category = new Category
-            {
-                Name = categoryDto.Name,
-                UserId = userId
-            };
-
             _context.Categories.Add(category);
             await _context.SaveChangesAsync();
 
             return category;
         }
 
-        public async Task<Category> DeleteCategoryAsync(string userId, int categoryId)
+        public async Task<bool> DeleteCategoryAsync(string userId, int categoryId)
         {
             var category = await _context.Categories
+                .AsNoTracking()
                 .FirstOrDefaultAsync(c => c.Id == categoryId &&  c.UserId == userId);
+
+            if (category == null)
+                return false;
 
             _context.Categories.Remove(category);
             await _context.SaveChangesAsync();
 
-            return category;
+            return true;
         }
     }
 }
