+++ AngulaToDo.Server/Repositories/TaskRepository.cs
@@ -2,6 +2,7 @@
 using AngulaToDo.Server.Data.Dtos;
 using AngulaToDo.Server.Models;
 using AngulaToDo.Server.Repositories.Interfaces;
+using AutoMapper;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -14,31 +15,26 @@ namespace AngulaToDo.Server.Repositories
     public class TaskRepository : ITaskRepository
     {
         private readonly ToDoContext _context;
-        public TaskRepository(ToDoContext context)
+        private readonly IMapper _mapper;
+        public TaskRepository(ToDoContext context, IMapper mapper)
         {
             _context = context;
+            _mapper = mapper;
         }
 
         public async Task<IEnumerable<Task>> GetAllAsync(string userId)
         {
             var tasks = await _context.Tasks
                 .Where(t => t.UserId == userId)
+                .AsNoTracking()
                 .ToListAsync();
 
             return tasks;
         }
 
-        public async Task<Task> CreateTaskAsync(string userId, TaskDto taskDto)
+        public async Task<Task> CreateTaskAsync(string userId, Task task)
         {
-            var task = new Task
-            {
-                UserId = userId,
-                Title = taskDto.Title,
-                IsImportant = taskDto.IsImportant,
-                Created = DateTime.Now,
-                DueDate = taskDto.DueDate,
-                CategoryId = taskDto.CategoryId
-            };
+            task.UserId = userId;
 
             _context.Tasks.Add(task);
             await _context.SaveChangesAsync();
@@ -49,34 +45,37 @@ namespace AngulaToDo.Server.Repositories
         public async Task<Task> GetTaskByIdAsync(string userId, int taskId)
         {
             var task = await _context.Tasks
+                .AsNoTracking()
                 .FirstOrDefaultAsync(t => t.UserId == userId && t.Id == taskId);
 
             return task;
         }
 
-        public async Task<Task> UpdateTaskAsync(string userId, int taskId, TaskDto taskDto)
+        public async Task<bool> UpdateTaskAsync(string userId, int taskId, TaskDto taskDto)
         {
             var task = await GetTaskByIdAsync(userId, taskId);
 
-            task.Title = taskDto.Title;
-            task.IsImportant = taskDto.IsImportant;
-            task.Completed = taskDto.Completed;
-            task.DueDate = taskDto.DueDate;
+            if (task == null)
+                return false;
+            task = _mapper.Map(taskDto, task);
 
             _context.Tasks.Update(task);
             await _context.SaveChangesAsync();
 
-            return task;
+            return true;
         }
 
-        public async Task<Task> DeleteTaskAsync(string userId, int taskId)
+        public async Task<bool> DeleteTaskAsync(string userId, int taskId)
         {
             var task = await GetTaskByIdAsync(userId, taskId);
+
+            if (task == null)
+                return false;
 
             _context.Tasks.Remove(task);
             await _context.SaveChangesAsync();
 
-            return task;
+            return true;
         }
     }
 }
